+++ examples/aiohttp/template.py
@@ -1,9 +1,9 @@
-
 from string import Template
 
 
 def render_graphiql():
-    return Template('''
+    return Template(
+        """
 <!DOCTYPE html>
 <html>
 <head>
@@ -116,10 +116,11 @@ def render_graphiql():
     );
   </script>
 </body>
-</html>''').substitute(
-        GRAPHIQL_VERSION='0.10.2',
-        SUBSCRIPTIONS_TRANSPORT_VERSION='0.7.0',
-        subscriptionsEndpoint='ws://localhost:8000/subscriptions',
+</html>"""
+    ).substitute(
+        GRAPHIQL_VERSION="0.10.2",
+        SUBSCRIPTIONS_TRANSPORT_VERSION="0.7.0",
+        subscriptionsEndpoint="ws://localhost:8000/subscriptions",
         # subscriptionsEndpoint='ws://localhost:5000/',
-        endpointURL='/graphql',
+        endpointURL="/graphql",
     )