+++ examples/django_subscriptions/django_subscriptions/schema.py
@@ -6,18 +6,19 @@ class Query(graphene.ObjectType):
     hello = graphene.String()
 
     def resolve_hello(self, info, **kwargs):
-        return 'world'
+        return "world"
+
 
 class Subscription(graphene.ObjectType):
 
     count_seconds = graphene.Int(up_to=graphene.Int())
 
-
     def resolve_count_seconds(root, info, up_to=5):
-        return Observable.interval(1000)\
-                         .map(lambda i: "{0}".format(i))\
-                         .take_while(lambda i: int(i) <= up_to)
-
+        return (
+            Observable.interval(1000)
+            .map(lambda i: "{0}".format(i))
+            .take_while(lambda i: int(i) <= up_to)
+        )
 
 
 schema = graphene.Schema(query=Query, subscription=Subscription)