+++ graphql_ws/__init__.py
@@ -3,8 +3,5 @@
 """Top-level package for GraphQL WS."""
 
 __author__ = """Syrus Akbary"""
-__email__ = 'me@syrusakbary.com'
-__version__ = '0.1.0'
-
-
-from .base import BaseConnectionContext, BaseSubscriptionServer
+__email__ = "me@syrusakbary.com"
+__version__ = "0.3.1"