+++ graphql_ws/observable_aiter.py
@@ -1,7 +1,7 @@
+from asyncio import Future
 
-from asyncio import Future, get_event_loop
-from rx.internal import extensionmethod
 from rx.core import Observable
+from rx.internal import extensionmethod
 
 
 async def __aiter__(self):
@@ -13,15 +13,11 @@ async def __aiter__(self):
             self.future = Future()
 
             self.disposable = source.materialize().subscribe(self.on_next)
-            # self.disposed = False
 
         def __aiter__(self):
             return self
 
         def dispose(self):
-            # self.future.cancel()
-            # self.disposed = True
-            # self.future.set_exception(StopAsyncIteration)
             self.disposable.dispose()
 
         def feeder(self):
@@ -30,11 +26,11 @@ async def __aiter__(self):
 
             notification = self.notifications.pop(0)
             kind = notification.kind
-            if kind == 'N':
+            if kind == "N":
                 self.future.set_result(notification.value)
-            if kind == 'E':
+            if kind == "E":
                 self.future.set_exception(notification.exception)
-            if kind == 'C':
+            if kind == "C":
                 self.future.set_exception(StopAsyncIteration)
 
         def on_next(self, notification):
@@ -42,8 +38,6 @@ async def __aiter__(self):
             self.feeder()
 
         async def __anext__(self):
-            # if self.disposed:
-            #     raise StopAsyncIteration
             self.feeder()
 
             value = await self.future
@@ -51,39 +45,6 @@ async def __aiter__(self):
             return value
 
     return AIterator()
-
-
-# def __aiter__(self, sentinel=None):
-#     loop = get_event_loop()
-#     future = [Future()]
-#     notifications = []
-
-#     def feeder():
-#         if not len(notifications) or future[0].done():
-#             return
-#         notification = notifications.pop(0)
-#         if notification.kind == "E":
-#             future[0].set_exception(notification.exception)
-#         elif notification.kind == "C":
-#             future[0].set_exception(StopIteration(sentinel))
-#         else:
-#             future[0].set_result(notification.value)
-
-#     def on_next(value):
-#         """Takes on_next values and appends them to the notification queue"""
-#         notifications.append(value)
-#         loop.call_soon(feeder)
-
-#     self.materialize().subscribe(on_next)
-
-#     @asyncio.coroutine
-#     def gen():
-#         """Generator producing futures"""
-#         loop.call_soon(feeder)
-#         future[0] = Future()
-#         return future[0]
-
-#     return gen
 
 
 def setup_observable_extension():