+++ HISTORY.rst
@@ -2,7 +2,25 @@
 History
 =======
 
-0.1.0 (2017-10-15)
-------------------
+0.3.0 (2018-09-25)
+==================
 
-* First release on PyPI.
+- Django channels (v1) server (with example app).
+
+- Websockets lib server (with example app).
+
+- Observable aiter server.
+
+
+0.2.0 (2017-11-18)
+==================
+
+- Add gevent subscription server.
+
+- Add example flask gevent app.
+
+
+0.1.0 (2017-10-15)
+==================
+
+- First release on PyPI.